+++ src/save_data/mass_effect_1_le/player.rs
@@ -12,21 +12,21 @@ use crate::save_data::{
 #[derive(Deserialize, Serialize, Clone, RawUi)]
 pub struct Player {
     pub is_female: bool,
-    localized_class_name: i32,
-    player_class: u8,
+    pub localized_class_name: i32,
+    pub player_class: Me1LeClass,
     pub level: i32,
     pub current_xp: f32,
     pub first_name: String,
     localized_last_name: i32,
     pub origin: Origin,
     pub notoriety: Notoriety,
-    specialization_bonus_id: i32,
+    pub specialization_bonus_id: i32,
     spectre_rank: u8,
     pub talent_points: i32,
     talent_pool_points: i32,
     mapped_talent: String,
     pub head_morph: Option<HeadMorph>,
-    simple_talents: Vec<SimpleTalent>,
+    pub simple_talents: Vec<SimpleTalent>,
     pub complex_talents: Vec<ComplexTalent>,
     pub inventory: Inventory,
     pub credits: i32,
@@ -35,7 +35,7 @@ pub struct Player {
     pub omnigel: f32,
     pub face_code: String,
     armor_overridden: bool,
-    auto_levelup_template_id: i32,
+    pub auto_levelup_template_id: i32,
     health_per_level: f32,
     stability: f32,
     race: u8,
@@ -63,6 +63,16 @@ pub struct Player {
     secondary_weapon: String,
 }
 
+#[derive(Deserialize, Serialize, Clone, RawUi)]
+pub enum Me1LeClass {
+    Soldier,
+    Engineer,
+    Adept,
+    Infiltrator,
+    Sentinel,
+    Vanguard,
+}
+
 #[rcize_fields]
 #[derive(Deserialize, Serialize, Clone, Default, Display, RawUi)]
 #[display(fmt = "{}", talent_id)]
@@ -75,9 +85,9 @@ pub struct SimpleTalent {
 #[derive(Deserialize, Serialize, Clone, Default, Display, RawUi)]
 #[display(fmt = "{}", talent_id)]
 pub struct ComplexTalent {
-    talent_id: i32,
+    pub talent_id: i32,
     pub current_rank: i32,
-    max_rank: i32,
+    pub max_rank: i32,
     level_offset: i32,
     levels_per_rank: i32,
     visual_order: i32,
@@ -123,20 +133,20 @@ pub struct Item {
     pub item_id: i32,
     pub item_level: ItemLevel,
     pub manufacturer_id: i32,
-    plot_conditional_id: i32,
-    new_item: bool,
+    pub plot_conditional_id: i32,
+    pub new_item: bool,
     junk: bool,
-    attached_mods: Vec<ItemMod>,
+    pub attached_mods: Vec<ItemMod>,
 }
 
 #[rcize_fields]
 #[derive(Deserialize, Serialize, Clone, Default, Display, RawUi)]
 #[display(fmt = "")]
-struct ItemMod {
-    item_id: i32,
-    item_level: ItemLevel,
-    manufacturer_id: i32,
-    plot_conditional_id: i32,
+pub struct ItemMod {
+    pub item_id: i32,
+    pub item_level: ItemLevel,
+    pub manufacturer_id: i32,
+    pub plot_conditional_id: i32,
 }
 
 #[rcize_fields]