+++ src/services/database.rs
@@ -3,16 +3,20 @@ use std::rc::Rc;
 use anyhow::{Context as AnyhowContext, Error, Result};
 use yew_agent::{Agent, AgentLink, Context, HandlerId};
 
-use crate::save_data::{
-    mass_effect_1::plot_db::Me1PlotDb, mass_effect_1_le::item_db::Me1ItemDb,
-    mass_effect_2::plot_db::Me2PlotDb, mass_effect_3::plot_db::Me3PlotDb, shared::plot::RawPlotDb,
+use crate::{
+    save_data::{
+        mass_effect_1::plot_db::Me1PlotDb, mass_effect_1_le::item_db::Me1ItemDb,
+        mass_effect_1_le::player_class_db::Me1LePlayerClassDb, mass_effect_2::plot_db::Me2PlotDb,
+        mass_effect_3::plot_db::Me3PlotDb, shared::plot::RawPlotDb,
+    },
+    services::rpc,
 };
-use crate::services::rpc;
 
 pub enum Type {
+    Me1LePlayerClasses,
     Me1Plot,
     Me1RawPlot,
-    Me1ItemDb,
+    Me1Items,
     Me2Plot,
     Me2RawPlot,
     Me3Plot,
@@ -20,9 +24,10 @@ pub enum Type {
 }
 
 pub enum Database {
+    Me1LePlayerClasses(Rc<Me1LePlayerClassDb>),
     Me1Plot(Rc<Me1PlotDb>),
     Me1RawPlot(Rc<RawPlotDb>),
-    Me1ItemDb(Rc<Me1ItemDb>),
+    Me1Items(Rc<Me1ItemDb>),
     Me2Plot(Rc<Me2PlotDb>),
     Me2RawPlot(Rc<RawPlotDb>),
     Me3Plot(Rc<Me3PlotDb>),
@@ -45,6 +50,7 @@ pub enum Response {
 
 #[derive(Default)]
 struct Databases {
+    me1_le_player_classes: Option<Rc<Me1LePlayerClassDb>>,
     me1_plot: Option<Rc<Me1PlotDb>>,
     me1_raw_plot: Option<Rc<RawPlotDb>>,
     me1_item_db: Option<Rc<Me1ItemDb>>,
@@ -73,13 +79,16 @@ impl Agent for DatabaseService {
         match msg {
             Msg::DatabaseLoaded(who, db) => {
                 match db {
+                    Database::Me1LePlayerClasses(ref db) => {
+                        self.dbs.me1_le_player_classes = Some(Rc::clone(db));
+                    }
                     Database::Me1Plot(ref db) => {
                         self.dbs.me1_plot = Some(Rc::clone(db));
                     }
                     Database::Me1RawPlot(ref db) => {
                         self.dbs.me1_raw_plot = Some(Rc::clone(db));
                     }
-                    Database::Me1ItemDb(ref db) => {
+                    Database::Me1Items(ref db) => {
                         self.dbs.me1_item_db = Some(Rc::clone(db));
                     }
                     Database::Me2Plot(ref db) => {
@@ -104,6 +113,15 @@ impl Agent for DatabaseService {
     fn handle_input(&mut self, msg: Self::Input, who: HandlerId) {
         match msg {
             Request::Database(db_type) => match db_type {
+                Type::Me1LePlayerClasses => match self.dbs.me1_le_player_classes {
+                    Some(ref db) => {
+                        self.respond_db(who, Database::Me1LePlayerClasses(Rc::clone(db)))
+                    }
+                    None => self.load_db(who, "databases/me1_le_player_class_db.ron", |response| {
+                        let db = ron::from_str(&response)?;
+                        Ok(Database::Me1LePlayerClasses(Rc::new(db)))
+                    }),
+                },
                 Type::Me1Plot => match self.dbs.me1_plot {
                     Some(ref db) => self.respond_db(who, Database::Me1Plot(Rc::clone(db))),
                     None => self.load_db(who, "databases/me1_plot_db.ron", |response| {
@@ -118,11 +136,11 @@ impl Agent for DatabaseService {
                         Ok(Database::Me1RawPlot(Rc::new(db)))
                     }),
                 },
-                Type::Me1ItemDb => match self.dbs.me1_item_db {
-                    Some(ref db) => self.respond_db(who, Database::Me1ItemDb(Rc::clone(db))),
+                Type::Me1Items => match self.dbs.me1_item_db {
+                    Some(ref db) => self.respond_db(who, Database::Me1Items(Rc::clone(db))),
                     None => self.load_db(who, "databases/me1_item_db.ron", |response| {
                         let db = ron::from_str(&response)?;
-                        Ok(Database::Me1ItemDb(Rc::new(db)))
+                        Ok(Database::Me1Items(Rc::new(db)))
                     }),
                 },
                 Type::Me2Plot => match self.dbs.me2_plot {